Re-instantiate GithubApiClient regularly
Our application typically crashes after some time (>1 day) when running
in production, similarly to what we experienced with the ApolloClient
some time ago. We hope that re-instantiating the client will solvee
this issue. If not, we will have to do some form of try-catch within
the coroutines, for instance resulting in a reboot.