Catch JsonSyntaxException when querying github api for workflow runs
We're experiencing the following exception from time to time:

Exception in thread "main" com.google.gson.JsonSyntaxException:
  java.lang.IllegalStateException: Expected BEGIN_OBJECT but was
                                   STRING at line 1 column 1 path

I can think of a couple different cases where the github API may
reply with a string rather than json, so it's not unreasonable
to handle this. However, we're not particularly interested in
doing anything other than keeping the prosess alive and hoping
the following interation is successful. In order to keep track
of how often this occurs, I added some logging.